feat(ed): export onRequest function
Export the `onRequest` function to make it accessible outside of the module. This allows other modules or files to import and use the function.

This change was made to enable reusability and modularity of the `onRequest` function. It may impact any code that relies on importing this module, as they will need to update their imports to reflect the change.