feat(utils): add cn_hostnames array
This commit adds an array named `cn_hostnames` to the `utils.js` file, which contains a list of Chinese hostnames. This array is used in the main code to select a random hostname for proxying requests to different domains. The addition of this array improves the flexibility and customization of the code when handling requests. No significant consequences or considerations arise from this change.